requirements.txt
aiohttp==3.9.1
motor==3.3.1
pymongo==4.6.1
websockets==12.0
grpcio==1.59.3
grpcio-tools==1.59.3
python-jose==3.3.0
bcrypt==4.0.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
aiohttp_cors
pydantic[email]
passlib